Follow the move in the docs, and put the Lex test beside Lex
Fifth preflight round. The move itself is clean: 159 lines byte-identical
after the one `crate::source_header::Lex::new` -> `Lex::new` rewrite the new
import covers, 20 tests before and 20 after with matching names, and cargo
runs 336 either side. What it found is the paperwork.

`step_text.rs`'s module doc still advertised "header `FILE_SCHEMA` detection"
as one of the things it provides. That is exactly what left, and it is the
line a reader is most likely to trust. Five cross-references pointed at the old
home, one of them in `packages/parser/test/source-header.test.ts` telling a
TypeScript reader where the Rust guard lives. None break a build; all send
someone to a file that no longer has the thing. On a branch whose subject is
two halves whose comments cited each other while disagreeing, leaving those is
not a nit.

`the_comment_scan_stays_linear_on_hostile_input` was orphaned: it constructs
`Lex` and asserts on `lex.searches`, and after the move `step_text.rs` does not
mention `Lex` at all. It moves to a new `source_header_tests.rs`, beside the
type it tests. That file exists for what is `pub(crate)` and so unreachable
from an integration test, and its doc says to look at
`tests/source_header_comments.rs` first.

Also: the shortened import in `step.rs` fits on one line now, so rustfmt wanted
it collapsed. That violation was introduced by the previous commit, so it is
mine to clean even though the crate is broadly fmt-dirty and no CI job runs
`cargo fmt`. The only change to `step.rs` on this branch is that import, so
every other diff rustfmt reports there is on a line nothing here touched.
Eleven blank lines left in the hole the moved tests came out of, and
`mod schema_detect;` landed out of alphabetical order.

One correction to the previous commit message, since it named the wrong gate as
evidence: `node scripts/check-module-size.mjs` is TypeScript-only and passed at
527e47a1 with `step_text.rs` at 420 lines. The gate that was actually red is
the Rust one, `rust/processing/tests/module_size_ratchet.rs`, exit 101. The
commit body said so, but the verification list put the JS gate next to the
claim as though it were the check that caught it.

Verified by exit code, each on its own: export tests 0 (336), cargo test
--workspace 0, clippy -D warnings 0.
