+++ rust/export/src/lib.rs
@@ -39,10 +39,10 @@ mod parquet_bos;
 mod rooms;
 pub mod rooted_type;
 mod schema_convert;
+mod schema_detect;
 mod schema_pad;
 pub use schema_pad::padded_type_universe;
 mod shades;
-mod schema_detect;
 pub mod source_header;
 mod step;
 mod step_cow;
