+++ rust/export/src/step_text_tests.rs
@@ -7,16 +7,6 @@
 
 use super::*;
 
-
-
-
-
-
-
-
-
-
-
 /// End-to-end scenario for the same seam through the real `export_step`
 /// path: a source file whose `FILE_SCHEMA` label carries a literal `\`,
 /// exported with no explicit target schema (so `step.rs:196` falls back to
@@ -262,48 +252,4 @@ fn property_synthesis_round_trips_apostrophe_and_backslash_per_spec() {
         original_pset_name,
         "raw written bytes {raw_quoted_body:?} must spec-un-escape back to the original"
     );
-}
-
-#[test]
-fn the_comment_scan_stays_linear_on_hostile_input() {
-    // The property, not the timing. Searching at every `/*` is quadratic: a
-    // failing search runs to the end of the buffer, the caller advances one
-    // byte, and the next `/*` repeats it. This side is handed whole uncapped
-    // files by `detect_schema`, so it is the worse half.
-    //
-    // One failure proves no closer exists at or after any later position, and
-    // the scan only moves forward, so at most one search can ever fail and none
-    // should happen after it. Asserting the count pins that without a
-    // wall-clock threshold that would flake on a loaded machine.
-    //
-    // `/*` repeated would OVERLAP into `*/` and terminate itself, so the opens
-    // are spaced. The TypeScript twin of this test caught that fixture bug.
-    let hostile = format!("HEADER;\n{}\nFILE_SCHEMA;", "/* ".repeat(1000));
-    assert!(!hostile.contains("*/"));
-    let mut lex = crate::source_header::Lex::new(hostile.as_bytes());
-    let mut i = 0;
-    while i < hostile.len() {
-        match lex.skip_lexical_at(i) {
-            Some(end) if end > i => i = end,
-            _ => i += 1,
-        }
-    }
-    assert_eq!(lex.searches, 1, "one failed search should silence the rest");
-
-    // The memo must not fire early: 500 real comments, none unterminated, so
-    // every search succeeds and each consumes a span the others do not.
-    let packed = format!("HEADER;\n{}FILE_SCHEMA;", "/*x*/".repeat(500));
-    let mut lex = crate::source_header::Lex::new(packed.as_bytes());
-    let (mut i, mut skipped) = (0usize, 0u32);
-    while i < packed.len() {
-        match lex.skip_lexical_at(i) {
-            Some(end) if end > i => {
-                skipped += 1;
-                i = end;
-            }
-            _ => i += 1,
-        }
-    }
-    assert_eq!(skipped, 500);
-    assert_eq!(lex.searches, 500);
 }
